+++ packages/asc-ui/src/components/Heading/HeadingStyle.ts
@@ -1,13 +1,7 @@
-import styled, { css } from 'styled-components'
+import styled from 'styled-components'
 import type { Props as TypographyProps } from '../Typography/TypographyStyle'
 import Typography from '../Typography'
 
 export type Props = TypographyProps
 
-export const HeaderStyleCSS = css`
-  margin-top: 0;
-`
-
-export default styled(Typography)<Props>`
-  ${HeaderStyleCSS}
-`
+export default styled(Typography)``