+++ packages/asc-ui/src/components/Typography/TypographyStyle.ts
@@ -6,16 +6,10 @@ import type { Theme } from '../../types'
 
 export interface Props {
   gutterBottom?: number
-  element?: Variant
-  fontSize?: number
   styleAs?: keyof Theme.TypographyElements
   as?: ElementType
   forwardedAs?: ElementType
   strong?: boolean
-  /**
-   * @deprecated Use your own custom style rules
-   */
-  color?: Theme.ColorType
 }
 
 export const defaultTypographyStyles: {
@@ -34,24 +28,15 @@ export default styled.p<Props>`
 
     return typeof key === 'string' && defaultTypographyStyles[key]
   }}
+
   margin: 0;
+
   ${getTypographyFromTheme()};
-  font-stretch: normal;
-  letter-spacing: normal;
+
   ${({ strong }) =>
     strong &&
     css`
       font-weight: 700;
       color: ${themeColor('tint', 'level7')};
-    `}
-  ${({ color }) =>
-    color &&
-    css`
-      color: ${themeColor(color)};
-    `}
-  ${({ fontSize }) =>
-    fontSize &&
-    css`
-      font-size: ${fontSize}px;
     `}
 `